backend/requirements.txt
fastapi>=0.104.0
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.1.0
python-multipart==0.0.9
pydantic[email]>=2.0.0